requirements.txt
altair==4.2.0
matplotlib==3.5.3
numerize==0.12
numpy==1.21.4
openpyxl==3.0.10
pandas==1.3.4
pdfplumber==0.5.28
Pillow==9.4.0
Plost==0.2.5
plotly==5.10.0
PyPDF2==3.0.1
seaborn==0.12.2
streamlit==1.20.0
streamlit_option_menu==0.3.2
XlsxWriter==3.0.3
streamlit==1.12.2
